action 3 ends step size 32 18 steps not intermediate_rewards not continue_until_ends only legal actions experiments memory errors

+++ codigo/it1/window_env_generator.py
@@ -193,7 +193,7 @@ class ImageWindowEnvGenerator(gym.Env):
         image_window_array = np.array(
             image_window_array)  # tf.keras.applications.mobilenet_v2.preprocess_input(image_window_resized)
         image_window_array = np.add(image_window_array, -128)
-        image_window_array = image_window_array / 128
+        image_window_array = np.float32(image_window_array / 128)
         return image_window_array
 
     def _get_predictions(self, image_window):
